File conversion utility. For now it only does unix2dos and dos2unix conversions.
There are samples in the javadoc, but the primary goal was to get something that could operate on an SVN
changeset to fix line endings and file formats automatically:
   --svn-update --dry-run --encoding=UTF-8 --line-ending=WINDOWS --match="(.java|.xml|.jsp|.properties)$" "w:\"
This will pick up all SVN-controlled files added or modified in the directory w:\ (unix paths accepted, too) and
show which need changing from Unix LF to Windows CR-LF line endings.
Remove the --dry-run to fix the files.
Add --no-backup if you trust the tool.

It is possible to find files by inclusive and exclusive regular expressions and doing directory recursion.