Partial fix for AWC-999: Content upload guesses character encoding and offers the user the chance to change it. TODO: Fix the "Modify Content Properties" to include the encoding as a changeable option.
Modified other entry points of content into the system.  All calls to ContentWriter.setEncoding("UTF-8") need some serious examination.
It is no longer necessary to assume anything about the encoding.  The worst case scenario is that we guess the encoding from the stream
without giving the user the chance to change it.  This works for most non-interactive scenarios like CIFS, WebDAV and FTP, now.