Point checkin. Got libssh2 to cross compile for linux mingw32 using libgcrypt, and got the curl library configured to link to it  (no testing yet).
This is supposed to work on linux, so ought to work on Windows; my assumption
is that the only thing really broken before was the libssh2 build.  Rather 
than fix the autotools configuration, the authors make an entirely separate
build for win32, and then hard-coded it for openssl, rather than libgcrypt,
and the ddl import names were wrong.  Aside:  I'd like to know if red-headed
stepchildren are in fact worse off than other kinds of stepchildren, and why.
Perhaps it depends on the hair color of the adopting family.

Anyway, the patch I've constructed does not fix the autoconf/automake config; 
it merely re-hacks the special purpose win32 build hack.  By the looks of it,
libssh2 is under rapid development now, so this arrangement may be "temporary".