Initial cut of first 2 WEF webscripts.
The startTemplate tag now renders a script tag referencing the /wef/bootstrap webscript, which in turn returns the contents of wef-boot.js (moved to the server side), sets up required constants and calls WEF.init().

The second webscript, /wef/resources renders a WEF.addResource() call for each resource configured in wef-context.xml, however, the object constructed is not complete and neither is the dependency tree in the spring config, so this webscript is not referenced at all yet.