Implemented temporary workflow form processor and associated test.
An item kind of "workflow" and itemId of "<workflow-def-name>" are now supported (only "jbpm$wf:adhoc" works though at the moment). Showing the form, selecting an assignee and submitting the form will start the ad hoc workflow and create a task for the selected assignee.

NOTE: Support for 'packageItems' is not present yet.