Fix for ALF-4647 (All sample workflows should be enabled OOTB) & ALF-4638 (The transition buttons on the edit task page sometimes change order) and a couple of other unreported issues found along the way.
The pooled and parallel workflow definitions are now enabled by default, all corresponding form configuration for the start tasks and other task types are also defined, some tweaking may be need after review from Linton. 

Fixed a potential NPE when transition id's are returned as null. Log warnings added to the task type config evaluator when the type can not be retrieved, which gives an indication as to why forms display all the default fields rather than the configured ones.

Changed process definitions to use a consistent ordering for the Approve and Reject transitions.