SVC60: Review the use of findAuthorities() within ScriptAuthorityService (ALF-9247)
ScriptGroup i.e. the "groups" root object now has a getGroups method that behaves similarly to People.getPeople wherein depending on the filter provided determines whether consistent results are returned. Any filter that can be done with a canned query i.e. startsWith queries or "*" will go to the the AuthorityService.getAuthorities method otherwise the lucene based AuthorityService.findAuthorities is used.

The UI also no longer sends a leading *, it it still supported but the user must add it themselves.